Pass config and app args to flyctl postgres attach
I spotted the following error on a review app:

+ flyctl postgres attach myapp-review-db --yes
Checking for existing attachments
Error: consumer app "myapp" already contains a secret named DATABASE_URL

Which suggests it was trying to attach the database to the `sevensage`
app instead of the review app.  This ensures we pass the review app
name and config just to be sure.